Moving test files as these are only for events
