+++ tests/db/event/delete.test.ts
@@ -1,8 +1,8 @@
 import { Db, MongoClient } from "mongodb";
-import { defaultAfterAll, defaultAfterEach, defaultBeforeAll, defaultBeforeEach, demoEventData } from "../utilities/setup";
+import { defaultAfterAll, defaultAfterEach, defaultBeforeAll, defaultBeforeEach, demoEventData } from "../../utilities/setup";
 import { BaseSchema } from "@uems/uemscommlib/build/BaseSchema";
 import Intentions = BaseSchema.Intentions;
-import { EventDatabase } from "../../src/database/type/impl/EventDatabaseInterface";
+import { EventDatabase } from "../../../src/database/type/impl/EventDatabaseInterface";
 
 const empty = <T extends Intentions>(intention: T): { msg_intention: T, msg_id: 0, status: 0, userID: string } => ({
     msg_intention: intention,
