+++ tests/db/event/query.test.ts
@@ -1,7 +1,7 @@
 import { Db, MongoClient } from "mongodb";
-import { defaultAfterAll, defaultAfterEach, defaultBeforeAll, defaultBeforeEach, demoEventData, empty } from "../utilities/setup";
+import { defaultAfterAll, defaultAfterEach, defaultBeforeAll, defaultBeforeEach, demoEventData, empty } from "../../utilities/setup";
 import { BaseSchema } from "@uems/uemscommlib/build/BaseSchema";
-import { EventDatabase } from "../../src/database/type/impl/EventDatabaseInterface";
+import { EventDatabase } from "../../../src/database/type/impl/EventDatabaseInterface";
 
 describe('create messages of states', () => {
     let client!: MongoClient;
@@ -60,6 +60,7 @@ describe('create messages of states', () => {
             stateID: 'teacher inside',
             entsID: 'embox dorm',
             start: 1612686824,
+            end: 1612751249,
             attendance: 185,
         });
